final routes

+++ server/src/controller/auth.ts
@@ -19,6 +19,9 @@ export const create: RequestHandler = async(req: CreateUser, res)=>{
    
 const {email, password, name} = req.body;
 
+const oldUser = await User.findOne({email})
+if(oldUser) return res.status(403).json({error: "Email already exist!"})
+
 const user = await User.create({name, email, password}); 
 // send verification email 
 const token = generateToken()
@@ -62,6 +65,8 @@ export const sendReverificationToken: RequestHandler = async(req, res)=>{
 
     const user = await User.findById(userId)
     if(!user) return res.status(403).json({error: "Invalid request!"})
+
+    if(user.verified) return res.status(422).json({error: "Your account is already verified!"})
 
    await EmailVerificationToken.findOneAndDelete({
       owner: userId
@@ -133,7 +138,7 @@ export const updatePassword: RequestHandler = async(req, res)=>{
   if(matched) return res.status(422).json({error: "The new password must be diffrent!"})
 
   user.password = password
-  await user.save()
+  await user.save() 
 
   await PasswordResetToken.findOneAndDelete({owner: user._id});
   // send success mail
